Move testing.T based Writer into iotest
